refactor(agent-task): split god file and add compact status summary (#4886, #4922, #4910, #4392, #4396, #4398)
Split the 2300-line src/commands/agent_task.rs god file into focused
sibling modules (args, auth, controller, run, status, tests) so the root
is a thin dispatcher under the structural thresholds (#4886). The
read-side handlers and the bounded stdin/secret + git-output reads now
live in separate files, clearing the unbounded-output-capture
false-positive on the root (#4922).

Factor the duplicated [verify, private_verify, private_gate_reveal]
field group into a shared VerifyGateOptions (serde) flattened into the
core options and a VerifyGateArgs (clap) flattened into the CLI args,
keeping flag names and serialized JSON shapes identical (#4910).

Make `agent-task status` emit a compact, recovery-first summary by
default: run id, state, totals, a per-task source table mapping task ids
back to issue URLs and prompt titles from plan.json (#4392), deduped
empty-uri-filtered refs (#4396), and a prominent risk-flags section
(suspicious-full-file-rewrite, secrets-redacted, missing-test-evidence)
surfaced before promotion/apply (#4398). The full verbose payload stays
available via `--full`.
