+++ homeboy.json
@@ -394,6 +394,7 @@
         "description": "Command adapter boundary violation: `{term}` appears at line {line} in {classification} context `{context}`",
         "exclude_path_contains": [
           "src/commands/agent_task.rs",
+          "src/commands/agent_task/",
           "src/commands/agent_task_summary.rs",
           "src/commands/api.rs",
           "src/commands/auth.rs",
@@ -969,7 +970,6 @@
         "field_patterns::src/cli_surface.rs::RepeatedFieldPattern",
         "field_patterns::src/command_contract/output.rs::RepeatedFieldPattern",
         "field_patterns::src/commands/adapter.rs::RepeatedFieldPattern",
-        "field_patterns::src/commands/agent_task.rs::RepeatedFieldPattern",
         "field_patterns::src/commands/agent_task_dispatch.rs::RepeatedFieldPattern",
         "field_patterns::src/commands/report/performance_digest.rs::RepeatedFieldPattern",
         "field_patterns::src/commands/runs.rs::RepeatedFieldPattern",
@@ -1041,7 +1041,6 @@
         "near-duplication::src/core/runner/connection.rs::NearDuplicate",
         "near-duplication::src/core/runner/connection_daemon.rs::NearDuplicate",
         "near-duplication::src/core/runner/worker.rs::NearDuplicate",
-        "output_capture::src/commands/agent_task.rs::UnboundedOutputCapture",
         "output_capture::src/core/agent_task_promotion.rs::UnboundedOutputCapture",
         "output_capture::src/core/deploy/execution.rs::UnboundedOutputCapture",
         "output_capture::src/core/lifecycle_store.rs::UnboundedOutputCapture",
@@ -1055,8 +1054,7 @@
         "remote_execution_preflight::src/core/runner/lab_workspaces.rs::RemoteExecutionPreflight",
         "remote_execution_preflight::src/core/runner/rig_materialization.rs::RemoteExecutionPreflight",
         "remote_execution_preflight::src/core/runner/worker.rs::RemoteExecutionPreflight",
-        "structural::src/commands/agent_task.rs::GodFile",
-        "structural::src/commands/agent_task.rs::HighItemCount",
+        "structural::src/commands/agent_task/tests.rs::HighItemCount",
         "structural::src/commands/bench/observation.rs::GodFile",
         "structural::src/commands/bench/tests.rs::HighItemCount",
         "structural::src/commands/component.rs::GodFile",
