+++ src/commands/agent_task/args/mod.rs
@@ -1,0 +1,370 @@
+//! Clap argument and subcommand definitions for the `agent-task` command tree.
+//!
+//! These types are the durable CLI contract surface. Keeping them in one
+//! sibling module lets the `agent_task` root stay a thin dispatcher over the
+//! handler modules (`auth`, `controller`, `run`, `status`, `review`).
+
+use clap::{Args, Subcommand};
+use homeboy::core::agent_tasks::gate::{AgentTaskGateRevealPolicy, VerifyGateOptions};
+
+use super::super::agent_task_dispatch::DispatchArgs;
+use super::review;
+
+mod auth;
+mod controller;
+
+pub use auth::{
+    AgentTaskAuthCommand, AgentTaskAuthMapEnvArgs, AgentTaskAuthMapKeychainBundleArgs,
+    AgentTaskAuthRemoveArgs, AgentTaskAuthSetConfigArgs, AgentTaskAuthSetKeychainArgs,
+    AgentTaskAuthSetKeychainBundleArgs, AgentTaskAuthStatusArgs,
+};
+pub use controller::{
+    AgentTaskControllerApplyEventArgs, AgentTaskControllerCommand, AgentTaskControllerFromSpecArgs,
+    AgentTaskControllerInitArgs, AgentTaskControllerMarkHumanReadyArgs,
+    AgentTaskControllerPlanArgs, AgentTaskControllerRunArgs, AgentTaskControllerRunNextArgs,
+    AgentTaskControllerStatusArgs,
+};
+
+#[derive(Args, Debug)]
+pub struct AgentTaskArgs {
+    #[command(subcommand)]
+    pub command: AgentTaskCommand,
+}
+
+#[derive(Subcommand, Debug)]
+pub enum AgentTaskCommand {
+    /// Sync a workspace if --runner is supplied, dispatch a repo-cooking task, and return the durable run id.
+    Cook(DispatchArgs),
+    /// Run a durable repo cook loop: dispatch, promote, verify, retry red gates, and finalize.
+    Loop(AgentTaskLoopArgs),
+    /// Build and dispatch common repo-cooking agent tasks without hand-authored provider JSON.
+    Dispatch(DispatchArgs),
+    /// Run an agent-task plan through extension-declared executor providers.
+    RunPlan(RunPlanArgs),
+    /// Execute a previously submitted durable agent-task run.
+    Run(StatusArgs),
+    /// Claim and execute the oldest queued durable agent-task run.
+    RunNext,
+    /// Persist an agent-task plan and return a durable run id without executing it.
+    Submit(SubmitArgs),
+    /// Read durable agent-task run status.
+    Status(StatusArgs),
+    /// List durable agent-task runs, newest first.
+    List,
+    /// List queued and running durable agent-task runs, newest first.
+    Active,
+    /// Show the latest durable agent-task run.
+    Latest,
+    /// Read durable agent-task run scheduler events.
+    Logs(StatusArgs),
+    /// List artifacts and evidence refs recorded for a completed run.
+    Artifacts(StatusArgs),
+    /// Mark a queued or stale-running durable agent-task run as cancelled.
+    Cancel(CancelArgs),
+    /// Resume a queued or stale-running durable run.
+    Resume(StatusArgs),
+    /// Submit a fresh durable run from an existing run's plan.
+    Retry(RetryArgs),
+    /// Build a durable aggregate review envelope from run state, logs, artifacts, and promotion hints.
+    Review(ReviewArgs),
+    /// Promote a completed generic patch artifact into a managed worktree.
+    Promote(PromoteArgs),
+    /// Finalize a green cook run into a review-ready pull request.
+    FinalizePr(FinalizePrArgs),
+    /// Convert deterministic gate results into a cook-loop retry or stop decision.
+    GateFeedback(GateFeedbackArgs),
+    /// List extension-declared agent-task executor providers and optional secret readiness.
+    Providers(ProvidersArgs),
+    /// Configure and inspect agent-task provider authentication secrets.
+    Auth(AgentTaskAuthArgs),
+    /// Create, inspect, and resume durable multi-agent loop controller state.
+    Controller(AgentTaskControllerArgs),
+}
+
+/// Shared deterministic verification gate flags. Flattened into every
+/// agent-task arg struct that runs `--verify` / `--private-verify` gates so the
+/// field group lives in exactly one place while CLI flag names stay identical.
+#[derive(Args, Debug, Clone)]
+pub struct VerifyGateArgs {
+    /// Deterministic verification command to run after each promotion.
+    #[arg(long = "verify", value_name = "COMMAND")]
+    pub verify: Vec<String>,
+
+    /// Private deterministic verification command to run after each promotion.
+    #[arg(long = "private-verify", value_name = "COMMAND")]
+    pub private_verify: Vec<String>,
+
+    /// Feedback policy for failed private gates.
+    #[arg(
+        long = "private-gate-reveal",
+        default_value = "summary-only",
+        value_name = "POLICY"
+    )]
+    pub private_gate_reveal: AgentTaskGateRevealPolicy,
+}
+
+impl From<VerifyGateArgs> for VerifyGateOptions {
+    fn from(args: VerifyGateArgs) -> Self {
+        VerifyGateOptions {
+            verify: args.verify,
+            private_verify: args.private_verify,
+            private_gate_reveal: args.private_gate_reveal,
+        }
+    }
+}
+
+#[derive(Args, Debug)]
+pub struct AgentTaskAuthArgs {
+    #[command(subcommand)]
+    pub command: AgentTaskAuthCommand,
+}
+
+#[derive(Args, Debug)]
+pub struct AgentTaskControllerArgs {
+    #[command(subcommand)]
+    pub command: AgentTaskControllerCommand,
+}
+
+#[derive(Args, Debug)]
+pub struct ProvidersArgs {
+    /// Secret environment variable name to check without exposing its value. Repeatable.
+    #[arg(long = "secret-env", value_name = "ENV")]
+    pub secret_env: Vec<String>,
+}
+
+#[derive(Args, Debug)]
+pub struct AgentTaskLoopArgs {
+    #[command(flatten)]
+    pub dispatch: DispatchArgs,
+
+    /// Repo-cooking goal. Alias for the dispatch prompt when --prompt is omitted.
+    #[arg(long, value_name = "TEXT")]
+    pub goal: Option<String>,
+
+    /// Target managed worktree handle where candidate patches are promoted.
+    #[arg(long, value_name = "HANDLE")]
+    pub to_worktree: String,
+
+    /// External workspace provider command. When omitted, HOMEBOY_AGENT_TASK_PROMOTION_COMMAND is used.
+    #[arg(long, value_name = "COMMAND")]
+    pub provider_command: Option<String>,
+
+    #[command(flatten)]
+    pub gates: VerifyGateArgs,
+
+    /// Maximum cook-loop gate attempts, including the first candidate.
+    #[arg(long = "max-attempts", default_value_t = 3, value_name = "N")]
+    pub max_attempts: u32,
+
+    /// Stop after the first green promotion without committing, pushing, or opening/updating a PR.
+    #[arg(long = "no-finalize")]
+    pub no_finalize: bool,
+
+    /// PR base branch used by finalization.
+    #[arg(long, default_value = "main", value_name = "BRANCH")]
+    pub base: String,
+
+    /// PR head branch. Defaults to the current branch in the promoted worktree.
+    #[arg(long, value_name = "BRANCH")]
+    pub head: Option<String>,
+
+    /// PR title. Defaults to a title derived from --repo or --task-url.
+    #[arg(long, value_name = "TEXT")]
+    pub title: Option<String>,
+
+    /// Commit message for the verified candidate changes.
+    #[arg(long, value_name = "TEXT")]
+    pub commit_message: Option<String>,
+
+    /// Protected branch that may not be finalized directly. Repeatable.
+    #[arg(long = "protected-branch", default_values_t = review::default_protected_branches(), value_name = "BRANCH")]
+    pub protected_branches: Vec<String>,
+
+    /// AI tool disclosure line for the PR body.
+    #[arg(long, default_value = "OpenCode (GPT-5.5)", value_name = "TEXT")]
+    pub ai_tool: String,
+
+    /// AI assistance scope for the PR body.
+    #[arg(
+        long,
+        default_value = "Drafted implementation and tests; Chris reviews and owns the change.",
+        value_name = "TEXT"
+    )]
+    pub ai_used_for: String,
+}
+
+#[derive(Args, Debug)]
+pub struct RunPlanArgs {
+    /// AgentTaskPlan JSON file, @file, or - for stdin.
+    #[arg(long, value_name = "PATH")]
+    pub plan: String,
+    /// Also persist the completed run lifecycle record under this id.
+    #[arg(long, value_name = "ID")]
+    pub record_run_id: Option<String>,
+}
+
+#[derive(Args, Debug)]
+pub struct SubmitArgs {
+    /// AgentTaskPlan JSON file, @file, or - for stdin.
+    #[arg(long, value_name = "PATH")]
+    pub plan: String,
+    /// Optional durable run id. Generated when omitted.
+    #[arg(long, value_name = "ID")]
+    pub run_id: Option<String>,
+}
+
+#[derive(Args, Debug)]
+pub struct StatusArgs {
+    /// Durable run id returned by `agent-task submit` or `agent-task run-plan --record-run-id`.
+    pub run_id: String,
+
+    /// Emit the full verbose payload (all artifact/evidence refs) instead of the
+    /// default compact, recovery-first summary.
+    #[arg(long)]
+    pub full: bool,
+}
+
+#[derive(Args, Debug)]
+pub struct RetryArgs {
+    /// Existing durable run id whose plan should be retried.
+    pub run_id: String,
+
+    /// Optional durable run id for the retry. Generated when omitted.
+    #[arg(long, value_name = "ID")]
+    pub new_run_id: Option<String>,
+
+    /// Execute the newly queued retry immediately.
+    #[arg(long)]
+    pub run: bool,
+}
+
+#[derive(Args, Debug)]
+pub struct CancelArgs {
+    /// Durable run id returned by `agent-task submit` or `agent-task run-plan --record-run-id`.
+    pub run_id: String,
+
+    /// Operator-visible reason stored on the durable run record.
+    #[arg(long, value_name = "TEXT")]
+    pub reason: Option<String>,
+}
+
+#[derive(Args, Debug)]
+pub struct ReviewArgs {
+    /// Durable run id returned by `agent-task submit`, `dispatch`, or `run-plan --record-run-id`.
+    pub run_id: String,
+
+    /// Target workspace handle to include in generated promotion commands.
+    #[arg(long, value_name = "HANDLE")]
+    pub to_worktree: Option<String>,
+
+    /// External workspace provider command to include in generated promotion commands.
+    #[arg(long, value_name = "COMMAND")]
+    pub provider_command: Option<String>,
+}
+
+#[derive(Args, Debug)]
+pub struct PromoteArgs {
+    /// AgentTaskOutcome or AgentTaskAggregate JSON file, @file, or - for stdin.
+    #[arg(value_name = "SOURCE")]
+    pub source: String,
+
+    /// Target workspace handle to apply into.
+    #[arg(long, value_name = "HANDLE")]
+    pub to_worktree: String,
+
+    /// External workspace provider command. When omitted, HOMEBOY_AGENT_TASK_PROMOTION_COMMAND is used.
+    #[arg(long, value_name = "COMMAND")]
+    pub provider_command: Option<String>,
+
+    /// Outcome task id to select when SOURCE is an aggregate.
+    #[arg(long, value_name = "TASK_ID")]
+    pub task_id: Option<String>,
+
+    /// Patch artifact id to select when the outcome contains multiple patches.
+    #[arg(long, value_name = "ARTIFACT_ID")]
+    pub artifact_id: Option<String>,
+
+    /// Validate and report the selected promotion without creating/applying.
+    #[arg(long)]
+    pub dry_run: bool,
+
+    #[command(flatten)]
+    pub gates: VerifyGateArgs,
+}
+
+#[derive(Args, Debug)]
+pub struct FinalizePrArgs {
+    /// Durable cook/agent-task run id to link in the PR body.
+    #[arg(long, value_name = "ID")]
+    pub run_id: String,
+
+    /// Verified candidate worktree path.
+    #[arg(long, value_name = "PATH")]
+    pub path: String,
+
+    /// PR base branch.
+    #[arg(long, default_value = "main", value_name = "BRANCH")]
+    pub base: String,
+
+    /// PR head branch. Defaults to the current branch in --path.
+    #[arg(long, value_name = "BRANCH")]
+    pub head: Option<String>,
+
+    /// PR title.
+    #[arg(long, value_name = "TEXT")]
+    pub title: String,
+
+    /// Commit message for the verified candidate changes.
+    #[arg(long, value_name = "TEXT")]
+    pub commit_message: String,
+
+    #[command(flatten)]
+    pub evidence: review::FinalizePrEvidenceArgs,
+
+    /// Green gate result as name=status or name=status:detail. Repeatable.
+    #[arg(long = "gate-result", value_name = "NAME=STATUS[:DETAIL]")]
+    pub gate_results: Vec<String>,
+
+    /// Changed file summary to include in output/PR body. Defaults to git status discovery.
+    #[arg(long = "changed-file", value_name = "PATH")]
+    pub changed_files: Vec<String>,
+
+    /// Protected branch that may not be finalized directly. Repeatable.
+    #[arg(long = "protected-branch", default_values_t = review::default_protected_branches(), value_name = "BRANCH")]
+    pub protected_branches: Vec<String>,
+
+    /// AI assistance scope for the PR body.
+    #[arg(
+        long,
+        default_value = "Drafted implementation and tests; Chris reviews and owns the change.",
+        value_name = "TEXT"
+    )]
+    pub ai_used_for: String,
+}
+
+#[derive(Args, Debug)]
+pub struct GateFeedbackArgs {
+    /// AgentTaskPromotionReport JSON file, @file, or - for stdin.
+    #[arg(long, value_name = "PATH")]
+    pub promotion: String,
+
+    /// Original AgentTaskRequest JSON file, @file, or - for stdin.
+    #[arg(long = "source-task", value_name = "PATH")]
+    pub source_task: String,
+
+    /// Current deterministic gate attempt number.
+    #[arg(long, default_value_t = 1, value_name = "N")]
+    pub attempt: u32,
+
+    /// Maximum deterministic gate attempts allowed for this cook loop.
+    #[arg(long = "max-attempts", default_value_t = 3, value_name = "N")]
+    pub max_attempts: u32,
+
+    /// Durable source run id to include in evidence refs.
+    #[arg(long = "source-run-id", value_name = "ID")]
+    pub source_run_id: Option<String>,
+
+    /// Current candidate diff/context as a string, @file, or - for stdin.
+    #[arg(long = "current-diff", value_name = "SPEC")]
+    pub current_diff: Option<String>,
+}
