+++ src/commands/agent_task/review.rs
@@ -194,9 +194,7 @@ pub(crate) fn promote_artifact(args: PromoteArgs) -> CmdResult<Value> {
         task_id: args.task_id,
         artifact_id: args.artifact_id,
         dry_run: args.dry_run,
-        verify: args.verify,
-        private_verify: args.private_verify,
-        private_gate_reveal: args.private_gate_reveal,
+        gates: args.gates.into(),
         provider_command: args.provider_command,
     })?;
     let exit_code = if report.status == AgentTaskPromotionStatus::GateFailed {
